Split score_items and score_ambitos
